+++ src/Umbraco.Cms.ManagementApi/DependencyInjection/MappingBuilderExtensions.cs
@@ -6,7 +6,6 @@ using Umbraco.Cms.ManagementApi.Mapping.Installer;
 using Umbraco.Cms.ManagementApi.Mapping.Languages;
 using Umbraco.Cms.ManagementApi.Mapping.Relation;
 using Umbraco.Cms.ManagementApi.Mapping.TrackedReferences;
-using Umbraco.New.Cms.Infrastructure.Persistence.Mappers;
 
 namespace Umbraco.Cms.ManagementApi.DependencyInjection;
 
@@ -17,7 +16,6 @@ public static class MappingBuilderExtensions
         builder.WithCollectionBuilder<MapDefinitionCollectionBuilder>()
             .Add<DictionaryViewModelMapDefinition>()
             .Add<TrackedReferenceViewModelsMapDefinition>()
-            .Add<RelationModelMapDefinition>()
             .Add<RelationViewModelsMapDefinition>()
             .Add<LanguageViewModelsMapDefinition>()
             .Add<InstallerViewModelsMapDefinition>()
